Add static assets (mostly icons) Add manifest for PWA Add api functionality for more complex playlist pulls Add UI functionality for playlist pulls

+++ adless/tools.py
@@ -3,12 +3,15 @@ from clilib.util.logging import Logging
 from adless.plex import media_exists
 from urllib.parse import parse_qs, urlparse
 from pytube import YouTube, Playlist
+# from pytube.exceptions import AgeRestrictedError, VideoUnavailable
+import pytube.exceptions as exceptions
 from pathlib import Path
 from hashlib import sha1
 import requests
 import ffmpeg
 import redis
 import json
+import re
 import os
 
 
@@ -32,19 +35,45 @@ def on_complete(stream, file_path):
     """
     print(f"Downloaded to {file_path}")
 
-def get_video(video_id: str, destination: str, video_subdir: bool = True, only_audio: bool = False, itag: int = None):
+def get_video(video_id: str, destination: str, video_subdir: bool = True, only_audio: bool = False, itag: int = None, filename: str = None):
     """
     download a single video
     """
     destination = Path(destination)
     video = YouTube(video_id, on_progress_callback=on_progress, on_complete_callback=on_complete)
     video_name = video.title
+    if filename:
+        video_name = filename
+    video_name = video_name.replace(",", "").replace("|", "").replace("/", "")
     logger = Logging(video.video_id).get_logger()
-    if video_subdir:
-        destination = destination.joinpath(video.title)
-    if not destination.exists():
-        destination.mkdir(parents=True)
-    logger.info(f"Downloading video: {video.title}")
+    logger.info(f"Downloading video: {video_name} ...")
+    if only_audio:
+        logger.info("only_audio is set to True, assuming this is music ...")
+        # Assume this is being downloaded as music, and attempt to get track information
+        logger.info("Getting track information ...")
+        metadata = {}
+        if video.description:
+            matches = re.findall(r"([A-Z][a-z]+): (.*)", video.description)
+            for match in matches:
+                key = match[0].lower()
+                if key in metadata:
+                    continue
+                metadata[key] = match[1]
+        metadata["title"] = video_name
+        metadata["artist"] = video.author
+        if "album" not in metadata:
+            metadata["album"] = "Unknown"
+
+    if only_audio:
+        destination = destination.joinpath(metadata["artist"]).joinpath(metadata["album"])
+        if not destination.exists():
+            destination.mkdir(parents=True)
+    else:
+        if video_subdir:
+            destination = destination.joinpath(video_name)
+        if not destination.exists():
+            destination.mkdir(parents=True)
+
     thumbnail = destination.joinpath(f"{video_name}.jpg")
     res = requests.get(video.thumbnail_url)
     with open(thumbnail, "wb") as f:
@@ -70,14 +99,15 @@ def get_video(video_id: str, destination: str, video_subdir: bool = True, only_a
         _v = ffmpeg.input(video_file)
         _a = ffmpeg.input(audio_file)
         ffmpeg.output(_v, _a, output_file, acodec='copy', vcodec='copy').run()
+        video_file.unlink()
+    if not only_audio:
         audio_file.unlink()
-    video_file.unlink()
     full_url = f"https://www.youtube.com/watch?v={video.video_id}"
     key_name = f"video:{sha1(full_url.encode('utf-8')).hexdigest()}"
     db.lpush("downloads", key_name)
     # video.download(destination=destination, resolution="720p")
 
-def get_playlist_videos(playlist_id: str, destination: str):
+def get_playlist_videos(playlist_id: str, destination: str, only_audio: bool = False):
     """
     download all videos in a playlist
     """
@@ -87,9 +117,9 @@ def get_playlist_videos(playlist_id: str, destination: str):
     playlist = Playlist(playlist_id)
     logger = Logging(video.video_id).get_logger()
     logger.info(f"Downloading playlist: {playlist.title}")
-    destination = destination.joinpath(playlist.title)
+    # destination = destination.joinpath(playlist.title)
     for i, video in enumerate(playlist.video_urls):
-        get_video(video, destination, video_subdir=False)
+        get_video(video, destination, only_audio=only_audio)
         # video.download(destination=destination, resolution="720p"
     full_url = f"https://www.youtube.com/playlist?list={playlist.playlist_id}"
     key_name = f"playlist:{sha1(full_url.encode('utf-8')).hexdigest()}"
@@ -131,6 +161,9 @@ def get_video_info(video_id: str):
     """
     Get video info from YouTube and cache it in Redis for 24 hours.
     """
+    parsed_url = urlparse(video_id)
+    parsed_args = parse_qs(parsed_url.query)
+    real_id = parsed_args["v"][0]
     hashed_url = sha1(video_id.encode("utf-8")).hexdigest()
     key_name = f"video:{hashed_url}"
     if db.exists(key_name):
@@ -139,20 +172,82 @@ def get_video_info(video_id: str):
             info["_keyname"] = key_name
             info["_downloaded"] = media_exists(info["title"])
             return info
-    video = YouTube(video_id)
+    try:
+        # print("Getting video info (%s) ..." % video_id)
+        video = YouTube(video_id)
+        # print("Got video info (%s) ..." % video_id)
+    except exceptions.AgeRestrictedError:
+        return {
+            "id": real_id,
+            "title": "Age Restricted Video",
+            "description": "Unable to download age-restricted videos at this time ...",
+            "thumbnail": None,
+            "author": "Unknown",
+            "length": 0,
+            "views": 0,
+            "rating": None,
+            "publish_date": 0,
+            "streams": [],
+            "_type": "video",
+            "_pull_time": int(datetime.now().timestamp()),
+            "_keyname": key_name,
+            "_downloaded": False,
+            "_unavailable": True
+        }
+    except Exception as e:
+        # print("Unable to get video info (%s): %s" % (video_id, e))
+        return {
+            "id": real_id,
+            "title": "Video Unavailable (%s)" % e,
+            "description": "Unable to download video at this time ...",
+            "thumbnail": None,
+            "author": "Unknown",
+            "length": 0,
+            "views": 0,
+            "rating": None,
+            "publish_date": 0,
+            "streams": [],
+            "_type": "video",
+            "_pull_time": int(datetime.now().timestamp()),
+            "_keyname": key_name,
+            "_downloaded": False,
+            "_unavailable": True
+        }
     streams = []
-    for stream in video.streams.filter(adaptive=True, mime_type="video/webm", only_video=True):
-        streams.append({
-            "itag": stream.itag,
-            "mime_type": stream.mime_type,
-            "resolution": stream.resolution,
-            "fps": stream.fps,
-            "bitrate": stream.bitrate,
-            "type": "video"
-        })
+    try:
+        for stream in video.streams.filter(adaptive=True, mime_type="video/webm", only_video=True):
+            streams.append({
+                "itag": stream.itag,
+                "mime_type": stream.mime_type,
+                "resolution": stream.resolution,
+                "fps": stream.fps,
+                "bitrate": stream.bitrate,
+                "type": "video"
+            })
+    except exceptions.AgeRestrictedError:
+        return {
+            "id": real_id,
+            "title": "Age Restricted Video",
+            "description": "Unable to download age-restricted videos at this time ...",
+            "thumbnail": None,
+            "author": "Unknown",
+            "length": 0,
+            "views": 0,
+            "rating": None,
+            "publish_date": 0,
+            "streams": [],
+            "_type": "video",
+            "_pull_time": int(datetime.now().timestamp()),
+            "_keyname": key_name,
+            "_downloaded": False,
+            "_unavailable": True
+        }
     info = {
         "id": video.video_id,
-        "title": video.title,
+        # Remove commas due to plex bug which causes incorrect 
+        # media to be returned when title contains comma.
+        "_keyname": key_name,
+        "title": video.title.replace(",", "").replace("|", "").replace("/", ""),
         "description": video.description,
         "thumbnail": video.thumbnail_url,
         "author": video.author,
@@ -167,8 +262,8 @@ def get_video_info(video_id: str):
     info["_pull_time"] = int(datetime.now().timestamp())
     info_dump = json.dumps(info)
     db.set(key_name, info_dump)
-    info["_keyname"] = key_name
     info["_downloaded"] = media_exists(info["title"])
+    info["_unavailable"] = False
     return info
 
 def unshorten(url: str):
